requirements.txt
alembic==1.10.2
anyio==3.6.2
asyncio==3.4.3
asyncpg==0.27.0
beautifulsoup4==4.12.0
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
databases==0.7.0
fastapi==0.95.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.2
pydantic==1.10.7
python-dotenv==0.21.1
requests==2.28.2
sniffio==1.3.0
soupsieve==2.4
SQLAlchemy==1.4.46
starlette==0.26.1
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
